[REST] MFA Handling

+++ discord_client_rest/examples/guild.rs
@@ -1,5 +1,6 @@
 use discord_client_rest::image::ImageType;
 use discord_client_rest::image::base64::encode_image;
+use discord_client_rest::mfa::{MfaRequiredError, MfaType};
 use discord_client_rest::rest::RestClient;
 use discord_client_structs::structs::guild::create::CreateGuildBuilder;
 
@@ -32,5 +33,39 @@ async fn main() {
 
     println!("Got {} logs", logs.audit_log_entries.len());
 
-    client.guild(Some(guild.id)).delete().await.unwrap();
+    let result = client.guild(Some(guild.id)).delete().await; // boxed result
+
+    if let Err(err) = &result {
+        if let Some(mfa_error) = err.downcast_ref::<MfaRequiredError>() {
+            let verification_request = &mfa_error.verification_request;
+            let ticket = verification_request.clone().ticket;
+            let mut found_totp_method = false;
+            for method in &verification_request.methods {
+                if let MfaType::Totp() = method.r#type {
+                    found_totp_method = true;
+
+                    println!("Enter your TOTP code:");
+                    let mut code = String::new();
+                    std::io::stdin().read_line(&mut code).unwrap();
+
+                    client
+                        .auth()
+                        .validate_mfa(code.trim().to_string(), MfaType::Totp(), ticket)
+                        .await
+                        .unwrap();
+
+                    client.guild(Some(guild.id)).delete().await.unwrap();
+
+                    break;
+                }
+            }
+
+            if !found_totp_method {
+                println!("No TOTP method found");
+                result.unwrap();
+            }
+        } else {
+            result.unwrap();
+        }
+    }
 }
